+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Correctly handle `var` in `--fix-function-typedefs` (#826).
 * Preserve leading indentation in fixed doc comments (#821).
 * Add `--verbose` to hide advanced options in `--help` output
+* Add `--verbose` to hide advanced options in `--help` output.
+* Split outer nested control flow elements (#869).
 * Always place a blank line after script tags (#782).
 
 # 1.3.3